refactor: move utils functions to class
