--- src/logging/greeting.js
@@ -1,6 +1,0 @@
-import { getArgValue } from "../utils/get-arg-value.utils.js"
-
-export const greeting = (key) => {
-    const userName = getArgValue(key)
-    console.log(`Welcome to the File Manager, ${userName}!`)
-}
