--- src/logging/show-current-directory.js
@@ -1,1 +1,0 @@
-export const showCurrentDirectory = (path) => console.log(`You are currently in ${path}`)
