--- src/utils/get-arg-value.utils.js
@@ -1,4 +1,0 @@
-export const getArgValue = key => {
-    const findedArg = process.argv.find((arg) => arg.startsWith(key)) ?? ''
-    return findedArg.slice(key.length + 1)
-}
